Created the Search functionality
I have created the search results page and route for the search to
work. Validation is not done yet. But the search work and renders the
data fine. The search will be performed on all the field of the main
document and the sub document which are of type string.